+++ packages/cli/src/commands/init/index.ts
@@ -2,9 +2,9 @@ import path from "path";
 import fs from "fs-extra";
 import chalk from "chalk";
 
-import logger from "../../utils/logger";
 import getChangesetBase from "../../utils/getChangesetBase";
 import { defaultWrittenConfig } from "@changesets/config";
+import { info, log, warn, error } from "@changesets/logger";
 
 const pkgPath = path.dirname(require.resolve("@changesets/cli/package.json"));
 
@@ -14,21 +14,21 @@ export default async function init(cwd: string) {
   if (fs.existsSync(changesetBase)) {
     if (!fs.existsSync(path.join(changesetBase, "config.json"))) {
       if (fs.existsSync(path.join(changesetBase, "config.js"))) {
-        logger.error(
+        error(
           "It looks like you're using the version 1 `.changeset/config.js` file"
         );
-        logger.error(
+        error(
           "The format of the config object has significantly changed in v2 as well"
         );
-        logger.error(
+        error(
           " - we thoroughly recommend looking at the changelog for this package for what has changed"
         );
-        logger.error(
+        error(
           "Changesets will write the defaults for the new config, remember to transfer your options into the new config at `.changeset/config.json`"
         );
       } else {
-        logger.error("It looks like you don't have a config file");
-        logger.info(
+        error("It looks like you don't have a config file");
+        info(
           "The default config file will be written at `.changeset/config.json`"
         );
       }
@@ -37,7 +37,7 @@ export default async function init(cwd: string) {
         JSON.stringify(defaultWrittenConfig, null, 2)
       );
     } else {
-      logger.warn(
+      warn(
         "It looks like you already have changesets initialized. You should be able to run changeset commands no problems."
       );
     }
@@ -48,17 +48,17 @@ export default async function init(cwd: string) {
       JSON.stringify(defaultWrittenConfig, null, 2)
     );
 
-    logger.log(
+    log(
       chalk`Thanks for choosing {green changesets} to help manage your versioning and publishing\n`
     );
-    logger.log("You should be set up to start using changesets now!\n");
+    log("You should be set up to start using changesets now!\n");
 
-    logger.info(
+    info(
       "We have added a `.changeset` folder, and a couple of files to help you out:"
     );
-    logger.info(
+    info(
       chalk`- {blue .changeset/README.md} contains information about using changesets`
     );
-    logger.info(chalk`- {blue .changeset/config.json} is our default config`);
+    info(chalk`- {blue .changeset/config.json} is our default config`);
   }
 }