+++ packages/cli/src/commands/publish/publishPackages.ts
@@ -3,7 +3,7 @@ import chalk from "chalk";
 import getWorkspaces from "../../utils/getWorkspaces";
 import { Workspace } from "get-workspaces";
 import * as npmUtils from "./npm-utils";
-import logger from "../../utils/logger";
+import { info, warn } from "@changesets/logger";
 import { TwoFactorState } from "../../utils/types";
 import { PreState } from "@changesets/types";
 
@@ -35,7 +35,7 @@ export default async function publishPackages({
   const unpublishedPackagesInfo = await getUnpublishedPackages(publicPackages);
 
   if (unpublishedPackagesInfo.length === 0) {
-    logger.warn("No unpublished packages to publish");
+    warn("No unpublished packages to publish");
   }
 
   workspacesByName;
@@ -75,7 +75,7 @@ async function publishAPackage(
   tag: string
 ) {
   const { name, version } = pkg.config;
-  logger.info(
+  info(
     `Publishing ${chalk.cyan(`"${name}"`)} at ${chalk.green(`"${version}"`)}`
   );
 
@@ -127,12 +127,12 @@ async function getUnpublishedPackages(packages: Array<Workspace>) {
       packagesToPublish.push(pkgInfo);
     } else if (semver.gt(localVersion, publishedVersion)) {
       packagesToPublish.push(pkgInfo);
-      logger.info(
+      info(
         `${name} is being published because our local version (${localVersion}) is ahead of npm's (${publishedVersion})`
       );
     } else if (semver.lt(localVersion, publishedVersion)) {
       // If the local version is behind npm, something is wrong, we warn here, and by not getting published later, it will fail
-      logger.warn(
+      warn(
         `${name} is not being published because version ${publishedVersion} is already published on npm and we are trying to publish version ${localVersion}`
       );
     }