+++ packages/cli/src/commands/version/version.test.ts
@@ -5,7 +5,8 @@ import path from "path";
 import fixturez from "fixturez";
 import versionCommand from "./index";
 import * as git from "@changesets/git";
-import logger from "../../utils/logger";
+import { warn } from "@changesets/logger";
+import { temporarilySilenceLogs } from "@changesets/test-utils";
 import writeChangeset from "../add/writeChangeset";
 import { NewChangeset, Config } from "@changesets/types";
 import { defaultConfig } from "@changesets/config";
@@ -28,7 +29,6 @@ const consoleError = console.error;
 
 jest.mock("../../utils/cli");
 jest.mock("@changesets/git");
-jest.mock("../../utils/logger");
 
 // @ts-ignore
 git.add.mockImplementation(() => Promise.resolve(true));
@@ -72,6 +72,7 @@ const getPkgJSON = (pkgName: string, calls: any) => {
 const writeEmptyChangeset = (cwd: string) => writeChangesets([], cwd);
 
 describe("running version in a simple project", () => {
+  temporarilySilenceLogs();
   let cwd: string;
 
   beforeEach(async () => {
@@ -89,7 +90,7 @@ describe("running version in a simple project", () => {
       await writeEmptyChangeset(cwd);
       await versionCommand(cwd, modifiedDefaultConfig);
       // @ts-ignore
-      const loggerWarnCalls = logger.warn.mock.calls;
+      const loggerWarnCalls = warn.mock.calls;
       expect(loggerWarnCalls.length).toEqual(1);
       expect(loggerWarnCalls[0][0]).toEqual(
         "No unreleased changesets found, exiting."